Fix 2 tests to work on python 3.8+ and on different timezones
There were 2 problems in tests.

The first one was in a test using `freezegun`, where we then used
`time_to_freeze.timestamp() - time.altzone` which was later compared
to `time.time()` in the code. However, `time.altzone` isn't always
the correct time offset, as it depends on daylight saving time.
Therefore, I changed it to use `time.time()` which is really what we
later use for comparison anyway.

The second one was a variable scoping bug. We created async tasks
with lambdas in a generator expression, in which the lambda used the
generator variable. This created a scoping issue (specifically in
python 3.8, probably because of async changes but I did not research
further) where 2 tasks would run with the same argument to the
lambda. I changed it so we pass the argument to `run_in_executor`
directly, fixing the scoping bug.