Add support for positional formatting arguments.
This is a fix for a very old issue (#97 in the previous repository).
Python's logging library allows us to send formatting arguments
either as a mapping or as positional arguments. Until now, aiologger
only supported a mapping as an argument, which rendered it
incompatible to python's original logging library, causing many
(including me) to not convert our logging to aiologger.

This commit adds support for the positional formatting arguments as
supported in python's logging library. The changed code which checks
for the validity of the arguments and "arranges" them as needed is
actually taken from the original logging library and integrated in
the library.

An additional test was added for positional formatting arguments, and
the test which checked the validity of arguments (which is no longer
required) was removed.